Introduce skipped tests against the server which introduce an endpoint enabling users to listen to the performance.
These tests use Capybara but they don't start a browser - they test the application in-memory, relying on the fact that it is a Sinatra app which complies with the Rack specification.
I've had to make some changes to the server to make this explicit so that we can use this testing strategy.
The build script now runs the backend tests too.