Spike - Connect AudioOutput to button on listen page.
This commit contains a lot of changes!

Firstly, what was known as `AudioOutput` is now `Power`. The metaphor has changed slightly. That's because I want to write tests against a 'component', which I'm defining as the marriage of some HTML and some JavaScript. I renamed the class to `Power` so I could give the name `AudioOutput` to the component. Looking at it now, I think I quite like the `Power` metaphor. At the moment the component only includes a 'power on button', so that seems to fit well.

The tests for the component are much more complicated than the tests for the class we had previously. However, you can see how the class tests have evolved - the remnants are still visible in the test named 'AudioOutput can be powered on', which has the same assertions but now operates at the level of interactions with HTML elements. I'm using Jest's "jsdom" environment to allow me to interact with the HTML, this is enabled by a comment at the top of the test file.

I also had to write a script that parses the erb file. There's a JavaScript function `compileErb` which I can use in my JavaScript tests. This function relies on a bash script I wrote (`render_erb.sh`) which takes a template as a string, some variables as a string and then uses Ruby to render the result. You don't really need to understand how this works, but I can explain more if you want to understand it better.

There's entry point to the JavaScript app is a file called `app.js` which at this point in time only `requires` and executes the AudioOutput component script. It uses common.js modules, so I'm using Parcel.js to produce a JavaScript file which the browser can run. Commonly, a single JavaScript file will be included on all of the pages of the website, so I've decided not to test for the presence of it - I'm happy to rely on the fact that we'll ensure the app.js bundle is always present on our pages.
I've added another step to the `build` script which produces the JavaScript bundle after all of the tests have passed.

In the `listen.erb` page the JavaScript bundle is added to the page via a <script> tag.
The AudioOutput erb file is also added to the page.

I replaced the test against the server for the button with a test that proves that the component HTML is present. The button is now tested in the JavaScript component tests, so I don't need to test for it twice. All that matters at the server level is that the component is included on the page.

I don't love this design. Actually, there are a number of things I don't like about it, especially the way the files are organised. We might move some of them around. However, I think this design follows a modular approach, the TDD has aided the design of the production code and I believe it's made it quite difficult to break accidentally. I'd be interested to know what ways you're able to break the application without making one of these tests fail 😁