Spike #2 - Connect AudioOutput to button on listen page.
This spike is much simpler than the previous one, but at a high level it achieves the same thing and once again there are tests (fewer though, since the implementationns a bit simpler). There are no components, no ERB parsing in JavaScript tests, no bash scripts, fewer modules. I think it's going to be harder to build on top of this, but it's probably easier to digest and work with at the beginning, so in that respect could be considered better (for our purposes).