Better naming of the server's 'streaming mode' configuration.
The instance of Kirill we'll run in production will need a 'StreamAlways' mode, which should be the default. We need to override it in the tests so that they don't try to stream from the server forever (we want the tests to pass/fail ASAP). The server's production code will need to use the streaming_mode to figure out whether to stream events to the client or not.
This implementation of the configuration is far from ideal as we need to explicitly notify the streaming mode when we've streamed an event, but until we have something better it should work. Our focus right now should be on making it work.